+++ lib/model/tempoh_jadual.dart
@@ -1,7 +1,16 @@
 enum TempohJadual {
   harini,
+
+  /// memberi respon untuk harini dan 7 hari berikutnya (jumlah respon ialah 8 hari). Terdapat pengecualian
+  /// di mana jika tempoh minggu itu akan bertukar tahun, pelayan akan mengeluarkan ralat
   minggu,
+
+  /// akan memberi respon untuk bulan semasa. Contohnya hari ini ialah 31 Disember 2023,
+  /// maka julat respon waktu solat ialah 1 Disember 2023 hingga 31 Disember 2023.
   bulan,
+
+  /// akan memberi respon untuk tahun semasa. Contohnya hari ini ialah 31 Disember 2023,
+  /// maka julat respon waktu solat ialah 1 Januari 2023 hingga 31 Disember 2023.
   tahun,
   durasi;
 }