+++ test/waktu_solat_malaysia_test.dart
@@ -2,12 +2,40 @@ import 'package:test/test.dart';
 import 'package:waktu_solat_malaysia/waktu_solat_malaysia.dart';
 
 void main() {
+  test('jumlah result dalam harini', () async {
+    final ws = await dapatkanJadualWaktuSolat(
+      ZonWaktuSolat.WLY01,
+      // tempohJadual: TempohJadual.harini,
+    );
+    print(ws?[0].toString());
+    expect(1, ws?.length);
+  });
+
   test('jumlah result dalam seminggu', () async {
     final ws = await dapatkanJadualWaktuSolat(
       ZonWaktuSolat.JHR02,
-      tempohJadual: TempohJadual.harini,
+      tempohJadual: TempohJadual.minggu,
     );
-    expect(1, ws?.length);
+    print(ws?[0].toString());
+    expect(8, ws?.length);
+  });
+
+  test('jumlah result dalam bulan', () async {
+    final ws = await dapatkanJadualWaktuSolat(
+      ZonWaktuSolat.JHR02,
+      tempohJadual: TempohJadual.bulan,
+    );
+    print(ws?[30].toString());
+    expect(31, ws?.length);
+  });
+
+  test('jumlah result dalam 1 tahun', () async {
+    final ws = await dapatkanJadualWaktuSolat(
+      ZonWaktuSolat.JHR02,
+      tempohJadual: TempohJadual.tahun,
+    );
+    print(ws?[0].toString());
+    expect(365, ws?.length);
   });
 
   test('jadual waktu solat berdurasi', () async {
@@ -17,13 +45,21 @@ void main() {
       ZonWaktuSolat.WLY01,
       tempohJadual: TempohJadual.durasi,
       mula: harini,
-      tamat: DateTime(
-        harini.year,
-        harini.month,
-        harini.day + 1,
-      ), // dapatkan tarikh esok
+      tamat: harini.add(Duration(days: 1)), // dapatkan tarikh esok
     );
-    expect(2, ws?.length);
+    expect(4, ws?.length);
+  });
+
+  test('jadual waktu solat berdurasi tetapi hanya harini', () async {
+    final harini = DateTime.now();
+
+    final ws = await dapatkanJadualWaktuSolat(
+      ZonWaktuSolat.WLY01,
+      tempohJadual: TempohJadual.durasi,
+      mula: harini,
+      tamat: harini,
+    );
+    expect(1, ws?.length);
   });
 
   test('tentukan zon', () {